Implement user-defined text color selection options
Description:
The goal of this addition was to include an additional flag so that users
can put in a hex or RGB value as a string and change the color of text output.

This is a new feature.

Note: only two additional packages have been added to the import.
Note: there is logic now that essentially chooses between the original
      text modification functions and the functions to parse user input.